refactor: participant 패키지 추가
