+++ src/main/java/model/BlackjackGame.java
@@ -6,6 +6,7 @@ import java.util.Map;
 import java.util.Set;
 import java.util.function.Function;
 import java.util.stream.Collectors;
+import model.participant.Participants;
 import model.result.BettingResults;
 import model.bet.ParticipantsBet;
 import model.cards.DealerCards;
